seperate inline code into own file
