+++ src/previewContainer.ts
@@ -2,19 +2,21 @@ import {PythonResult} from "arepl-backend"
 import * as vscode from "vscode"
 import PythonPanelPreview from "./pythonPanelPreview"
 import Reporter from "./telemetry"
-import Utilities from "./utilities"
 import {settings} from "./settings"
+import { pythonInlinePreview as PythonInlinePreview } from "./pythonInlinePreview"
 
 /**
  * logic wrapper around html preview doc
  */
 export class PreviewContainer{
     public printResults: string[];
+    pythonInlinePreview:PythonInlinePreview
     public errorDecorationType: vscode.TextEditorDecorationType
     private vars: {}
 
-    constructor(private reporter: Reporter, context: vscode.ExtensionContext, htmlUpdateFrequency=50, private pythonPreview?: PythonPanelPreview){
-        if(!this.pythonPreview) this.pythonPreview = new PythonPanelPreview(context, htmlUpdateFrequency);
+    constructor(private reporter: Reporter, context: vscode.ExtensionContext, htmlUpdateFrequency=50, private pythonPanelPreview?: PythonPanelPreview){
+        if(!this.pythonPanelPreview) this.pythonPanelPreview = new PythonPanelPreview(context, htmlUpdateFrequency)
+        this.pythonInlinePreview = new PythonInlinePreview(reporter, context)
         this.errorDecorationType = vscode.window.createTextEditorDecorationType({
             gutterIconPath: context.asAbsolutePath('media/red.jpg')
         })
@@ -22,7 +24,7 @@ export class PreviewContainer{
 
     public start(){
         this.clearStoredData()
-        return this.pythonPreview.start()
+        return this.pythonPanelPreview.start()
     }
 
     /**
@@ -51,7 +53,7 @@ export class PreviewContainer{
             }
             else{
                 // exec time is the 'truest' time that user cares about
-                this.pythonPreview.updateTime(pythonResults.execTime);
+                this.pythonPanelPreview.updateTime(pythonResults.execTime);
             }
 
             this.vars = {...this.vars, ...pythonResults.userVariables}
@@ -67,7 +69,7 @@ export class PreviewContainer{
             // So only update vars if there's not a syntax error
             // this is because it's annoying to user if they have a syntax error and all their variables dissapear
             if(!syntaxError){
-                this.pythonPreview.updateVars(this.vars)
+                this.pythonPanelPreview.updateVars(this.vars)
             }
 
             if(pythonResults.internalError){
@@ -84,15 +86,15 @@ export class PreviewContainer{
                 pythonResults.userErrorMsg = pythonResults.internalError
             }
 
-            if(this.printResults.length == 0) this.pythonPreview.clearPrint()
+            if(this.printResults.length == 0) this.pythonPanelPreview.clearPrint()
 
             this.updateError(pythonResults.userErrorMsg)
             if(settings().get('inlineResults')){
-                this.updateErrorGutterIcons(pythonResults.userErrorMsg)
+                this.pythonInlinePreview.updateErrorGutterIcons(pythonResults.userErrorMsg)
             }
 
-            this.pythonPreview.injectCustomCSS(settings().get('customCSS'))
-            this.pythonPreview.throttledUpdate()
+            this.pythonPanelPreview.injectCustomCSS(settings().get('customCSS'))
+            this.pythonPanelPreview.throttledUpdate()
 
             if(pythonResults.done) this.clearStoredData()
         } catch (error) {
@@ -121,72 +123,18 @@ export class PreviewContainer{
 
     public handlePrint(pythonResults: string){
         this.printResults.push(pythonResults);
-        this.pythonPreview.handlePrint(this.printResults.join('\n'))
+        this.pythonPanelPreview.handlePrint(this.printResults.join('\n'))
     }
 
     /**
      * @param refresh if true updates page immediately.  otherwise error will show up whenever updateContent is called
      */
     public updateError(err: string, refresh=false){
-        this.pythonPreview.updateError(err, refresh)
+        this.pythonPanelPreview.updateError(err, refresh)
     }
 
     public displayProcessError(err: string){
-        this.pythonPreview.displayProcessError(err)
-    }
-
-    /**
-     * sets gutter icons in sidebar. Safe - catches and logs any exceptions
-     */
-    private updateErrorGutterIcons(error: string){
-        try {
-            const errLineNums = this.getLineNumsFromPythonTrace(error)
-            
-            let decorations = errLineNums.map((num)=>{
-                const lineNum = num-1 // python trace uses 1-based indexing but vscode lines start at 0
-                const range = new vscode.Range(lineNum, 0, lineNum, 0)
-                return {range} as vscode.DecorationOptions
-            })
-            
-            if(vscode.window.activeTextEditor){
-                vscode.window.activeTextEditor.setDecorations(this.errorDecorationType, decorations)
-            }
-
-        } catch (error) {
-            if(error instanceof Error){
-                this.reporter.sendError(error)
-            }
-            else{
-                this.reporter.sendError(new Error(error))
-            }
-        }
-    }
-
-    /**
-     * returns line numbers for each error in the stack trace
-     * @param error a python stacktrace
-     */
-    private getLineNumsFromPythonTrace(error: string){
-            /* this regex will get the line number of each error. A error might look like this:
-            
-            Traceback (most recent call last):
-            line 4, in <module>
-            line 2, in foo
-            TypeError: unsupported operand type(s) for +: 'int' and 'str'
-            
-            The regex will not get line numbers in different files. Those have different format:
-            File "filePath", line 394, in func
-            */
-           const lineNumRegex = /^ *line (\d+), in /gm
-           let errLineNums: number[] = []
-           let match: RegExpExecArray
-           
-           while(match = lineNumRegex.exec(error)){
-               const matchCaptureGroup = match[1]
-               errLineNums.push(parseInt(matchCaptureGroup))
-           }
-
-           return errLineNums
+        this.pythonPanelPreview.displayProcessError(err)
     }
 
     /**
@@ -208,6 +156,6 @@ export class PreviewContainer{
     }
 
     get onDidChange(): vscode.Event<vscode.Uri> {
-        return this.pythonPreview.onDidChange
+        return this.pythonPanelPreview.onDidChange
     }
 }
